+++ core/shared/src/main/scala-2/me/shadaj/scalapy/py/AnyRawValue.scala
@@ -1,0 +1,15 @@
+package me.shadaj.scalapy.py
+
+import me.shadaj.scalapy.interpreter.PyValue
+
+trait AnyRawValue {
+  private[scalapy] val rawValue: PyValue
+}
+
+trait AnyPopulateWith {
+  def populateWith(v: PyValue): Any = {
+    new Any {
+      val rawValue = v
+    }
+  }
+}