+++ core/shared/src/main/scala/me/shadaj/scalapy/py/package.scala
@@ -1,6 +1,7 @@
 package me.shadaj.scalapy
 
 import scala.collection.mutable
+import scala.language.implicitConversions
 
 import scala.concurrent.Future
 
@@ -9,14 +10,13 @@ import me.shadaj.scalapy.readwrite.{Reader, Writer}
 import scala.collection.mutable.Queue
 import me.shadaj.scalapy.interpreter.Platform
 
-package object py {
+package object py extends PyMacros {
   def module(name: String) = Module(name)
   def module(name: String, subname: String) = Module(name, subname)
 
-  @py.native trait None extends Any
-  val None = Any.populateWith(CPythonInterpreter.noneValue).as[None]
+  val None = Any.populateWith(CPythonInterpreter.noneValue).as[PyNone.None]
 
-  type NoneOr[T] = None | T
+  type NoneOr[T] = PyNone.None | T
 
   def `with`[T <: py.Any, O](ref: T)(withValue: T => O): O = {
     ref.as[Dynamic](Reader.facadeReader[Dynamic](FacadeCreator.getCreator[Dynamic])).__enter__()
@@ -44,13 +44,13 @@ package object py {
     def convertProxy(v: T): PyValue
   }
 
-  implicit def seqConvertableSeqElem[T, C <% scala.collection.Seq[T]](implicit elemConvertable: ConvertableToSeqElem[T]): ConvertableToSeqElem[C] = new ConvertableToSeqElem[C] {
+  implicit def seqConvertableSeqElem[T, C](implicit ev: C => scala.collection.Seq[T], elemConvertable: ConvertableToSeqElem[T]): ConvertableToSeqElem[C] = new ConvertableToSeqElem[C] {
     def convertCopy(v: C): Platform.Pointer = {
-      CPythonInterpreter.createListCopy(v, elemConvertable.convertCopy)
+      CPythonInterpreter.createListCopy(ev(v), elemConvertable.convertCopy)
     }
 
     def convertProxy(v: C): PyValue = {
-      CPythonInterpreter.createListProxy(v, elemConvertable.convertProxy)
+      CPythonInterpreter.createListProxy(ev(v), elemConvertable.convertProxy)
     }
   }
 
@@ -59,7 +59,7 @@ package object py {
     def convertProxy(v: T): PyValue = writer.write(v)
   }
 
-  implicit class SeqConverters[T, C <% scala.collection.Seq[T]](seq: C) {
+  implicit class SeqConverters[T, C](seq: C)(implicit ev: C => scala.collection.Seq[T]) {
     def toPythonCopy(implicit elemWriter: ConvertableToSeqElem[T]): Any = {
       Any.populateWith(PyValue.fromNew(implicitly[ConvertableToSeqElem[C]].convertCopy(seq)))
     }
@@ -105,8 +105,4 @@ package object py {
       }
     }
   }
-
-  import scala.language.experimental.macros
-  def native[T]: T = macro FacadeImpl.native_impl[T]
-  def nativeNamed[T]: T = macro FacadeImpl.native_named_impl[T]
 }