+++ core/shared/src/test/scala-3/me/shadaj/scalapy/py/SpecialSyntaxTest.scala
@@ -1,0 +1,5 @@
+package me.shadaj.scalapy.py
+
+import org.scalatest.funsuite.AnyFunSuite
+
+trait CrossScalaSpecialSyntaxTest extends AnyFunSuite