+++ core/shared/src/test/scala/me/shadaj/scalapy/py/MethodCallingTest.scala
@@ -4,10 +4,6 @@ import me.shadaj.scalapy.interpreter
 
 import org.scalatest.funsuite.AnyFunSuite
 
-@native trait StringObjectFacade extends Object {
-  def replace(old: String, newValue: String): String = native
-}
-
 class MethodCallingTest extends AnyFunSuite {
   test("Can access global values") {
     local {