+++ coreMacros/src/main/scala-3/me/shadaj/scalapy/interpreter/ScalaNativeHelpers.scala
@@ -1,0 +1,5 @@
+package me.shadaj.scalapy.interpreter
+
+object ScalaNativeHelpers {
+  def inlineFnPtr2 = ???
+}